third_party/blink/web_tests/svg/hixie/error/013-expected.txt

CONSOLE ERROR: Error: <svg> attribute width: A negative value is not valid. ("-100")
